Factor ImmutableArray<T> into many files - Break immutable array (an enumerable, immutable, array) into a core component - And a bunch of algorithm components which represents algorithms on the array - And a bunch of components which represent creation of new related immutable arrays - The purpose of this work is to make it possible to source include the core of immutable array into a component which cannot depend on high level apis such as LINQ, SortedDictionary, etc. (All api use must be satisfied by System.Private.CoreLib). The work to add this other component (System.Private.Reflection.Metadata.Ecma335 will be delivered via a separate PR)
